requirements.txt
pyttsx3
SpeechRecognition
pyaudio
requests
PyYAML
pyqt5
